requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
PyJWT==2.8.0
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0